requirements.txt
python-telegram-bot==12.8
opencv-python==4.3.0.36
python-dotenv==0.15.0